requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-rest-swagger>=2.2.0,<2.3.0
drf-yasg==1.21.4
django-cors-headers>=3.11.0,<3.14.0